+++ src/rpc/methods/wallet.rs
@@ -190,10 +190,8 @@ impl RpcMethod<1> for WalletSetDefault {
         _: &http::Extensions,
     ) -> Result<Self::Ok, ServerError> {
         let mut keystore = ctx.keystore.write();
-        let addr_string = format!("wallet-{address}");
-        let key_info = keystore.get(&addr_string)?;
-        keystore.remove("default")?; // This line should unregister current default key then continue
-        keystore.put("default", key_info)?;
+        let key_info = crate::key_management::try_find(&address, &keystore)?;
+        keystore.set_default(key_info)?;
         Ok(())
     }
 }